Move the model generator functions to their own module
